Extract out distribution solvers
